requirements.txt
discord.py[voice]==1.7.3
yt-dlp==2022.8.8
asyncio==3.4.3